gcc: Add f95 -> gfortran symlink
Compiling R from source requires f95 as the program name for a Fortran 77 compiler. My understanding is that it can simply point to gfortran. Correct me if I'm wrong. This PR is entirely based on the existing line that symlinks `"gfortran-#{version_suffix}"` to `gfortran`. Let me know if this isn't the proper way of symlinking.